Small Improvements
Improvements:
- Default structure-mode is now local instead of server.
- It will now show if someone gets an extra turn.
- Extra turns are now limited to three in one streak.
- Added some translations.
- Moved from LibPG events to PaperMC ones.
- Some other small stuff.
- Fixed replays.
- Warps now actually have proper usage limitations even in easy difficulty.
- Cleaned up some code to be more Kotlin-like.
- Fixed the plugin breaking if no database exists.
- You no longer need to use Ctrl+Q to sell item stacks with more than one item.
- Player health is now the game's starting health instead of just 0.0.
- The actionbar is now actually displayed.
- Adjusted some item prices.

+++ src/main/kotlin/net/hectus/neobb/game/mode/PersonGame.kt
@@ -97,7 +97,7 @@ class PersonGame(world: World, bukkitPlayers: List<Player>, difficulty: GameDiff
                 Component.text("💎", Colors.PERSON_4).append(locale.component("scoreboard.rank", color = Colors.PERSON_0))
                     .append(Component.text("???")),
                 Component.text("⚔", Colors.PERSON_4).append(locale.component("scoreboard.elo", color = Colors.PERSON_0))
-                    .append(Component.text(player.databaseInfo.elo, Colors.PERSON_3)),
+                    .append(Component.text(player.databaseInfo.elo.toInt(), Colors.PERSON_3)),
                 Component.empty(),
                 Component.empty(),
                 Component.text("mc", Colors.PERSON_1).append(Component.text(".hectus", Colors.PERSON_2)).append(Component.text(".net", Colors.PERSON_3))
